aprsreceiver/requirements.txt
requests
fastapi
fastapi-utils[all]
uvicorn
influxdb-client
pyserial
